+++ code-convenience.js
@@ -6,6 +6,14 @@ function getPossibleBoolean(variable, _property) {
     return false
 }
 
+function getPossibleString(variable, _property) {
+    if (_property in variable) {
+        return variable[_property].get_string()[0]
+    }
+
+    return null
+}
+
 function startsWith(text, beginning) {
     return text.substr(0, beginning.length) == beginning
 }
@@ -23,4 +31,4 @@ function setIntersection(set, otherSet) {
     return new Set([...set].filter(item => otherSet.has(item)))
 }
 
-export { areArraysEqual, devLog, getPossibleBoolean, setIntersection, startsWith }
+export { areArraysEqual, devLog, getPossibleBoolean, getPossibleString, setIntersection, startsWith }